pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt==0.9",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]==0.115.6",
    "ipykernel==6.30.0",
    "jinja2==3.1.4",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-google-genai==2.1.8",
    "langchain-groq==0.3.6",
    "langchain-openai==0.2.10",
    "langsmith>=0.4.35",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "pypdf==3.14.0",
    "pytest>=8.4.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "sentence-transformers>=5.1.1",
    "structlog>=25.4.0",
    "uvicorn==0.32.1",
]
